Initial work to get webpack working
