+++ package.json
@@ -1,0 +1,47 @@
+{
+  "name": "stegaphoto",
+  "version": "1.0.0",
+  "description": "Simple website to hide files within images.",
+  "scripts": {
+    "start": "webpack-dev-server",
+    "build": "webpack"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/gregives/StegaPhoto.git"
+  },
+  "keywords": [
+    "javascript",
+    "steganography",
+    "subtlecrypto",
+    "web-workers",
+    "jszip",
+    "bootstrap4",
+    "stegaphoto",
+    "encryption"
+  ],
+  "author": "Greg Ives",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/gregives/StegaPhoto/issues"
+  },
+  "homepage": "https://github.com/gregives/StegaPhoto#readme",
+  "devDependencies": {
+    "@babel/cli": "^7.1.2",
+    "@babel/core": "^7.1.2",
+    "@babel/preset-env": "^7.1.0",
+    "webpack": "^4.20.2",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.1.9",
+    "worker-loader": "^2.0.0"
+  },
+  "dependencies": {
+    "babel-loader": "^8.0.4",
+    "bootstrap": "^4.1.3",
+    "html-webpack-plugin": "^3.2.0",
+    "jquery": "^3.3.1",
+    "jszip": "^3.1.5",
+    "popper.js": "^1.14.4",
+    "script-ext-html-webpack-plugin": "^2.0.1"
+  }
+}
